Added IJsonWritable interface
JsonNetSerializer no longer explicitly checks for a PersistentResponse to decide
whether to perform special serialization logic and instead will execute custom
serialization for any class that implements the IJsonWritable interface.
